requirements.txt
tensorflow-gpu==1.14
ffmpeg-python
streamlink
opencv-python
numpy